+++ lib/http.go
@@ -14,6 +14,7 @@ func (s *Server) ClearCookie(w http.ResponseWriter) {
 		Expires:  time.Now().Add(-1 * time.Hour),
 		MaxAge:   -1,
 		SameSite: http.SameSiteLaxMode,
+		Domain:   s.opts.CookieDomain,
 	})
 }